solved warning for 64bit arch

+++ SimpleTaskManager/SyncService/Operations/ReorderTaskOperation.h
@@ -10,8 +10,8 @@
 @interface ReorderTaskOperation : SingleOperation
 
 @property(nonatomic, copy) NSString *taskUid;
-@property(nonatomic) int targetIndex;
+@property(nonatomic) NSUInteger targetIndex;
 
-- (instancetype)initWithTaskUid:(NSString *)taskUid targetIndex:(int)targetIndex;
+- (instancetype)initWithTaskUid:(NSString *)taskUid targetIndex:(NSUInteger)targetIndex;
 
 @end
